Editei de double pra int

+++ matheus-586/aula-4/ex-7/Program.cs
@@ -1,16 +1,16 @@
 ﻿// - Escreva um programa que leia três notas de um aluno, calcule a média e exiba uma mensagem indicando se ele foi aprovado (média ≥ 7), está em recuperação (5 ≤ média < 7) ou foi reprovado (média < 5). Usar funções.
-double notaDoAluno()
+int notaDoAluno()
 {
     Console.WriteLine("escreva a nota do aluno: ");
-    return double.Parse(Console.ReadLine());
+    return int.Parse(Console.ReadLine());
 }
 
-double nota1 = notaDoAluno();
-double nota2 = notaDoAluno();
-double nota3 = notaDoAluno();
+int nota1 = notaDoAluno();
+int nota2 = notaDoAluno();
+int nota3 = notaDoAluno();
 
-double somatorio = nota1 + nota2 + nota3;
-double media = somatorio / 3;
+int somatorio = nota1 + nota2 + nota3;
+int media = somatorio / 3;
 
 if (media >= 7){
     Console.WriteLine("Você foi aprovado ");
